refactor: Nest tags tools under src/tools/tags/.
Move top-level tags MCP tools (and tests) into a resource folder per writing-mcp-tools conventions.
