+++ src/tools/index.ts
@@ -31,13 +31,12 @@ import "./list-cost-services";
 import "./list-folders";
 import "./list-provider-resources";
 import "./list-report-notifications";
-import "./list-tag-values";
-import "./list-tags";
 import "./network-flow-reports";
 import "./recommendation-views";
 import "./recommendations";
 import "./resource-reports";
 import "./submit-user-feedback";
+import "./tags";
 import "./update-anomaly";
 import "./update-folder";
 import "./update-report-notification";
