+++ src/tools/tags/index.ts
@@ -1,0 +1,2 @@
+import "./list-tag-values";
+import "./list-tags";
