+++ src/tools/tags/list-tag-values.ts
@@ -1,9 +1,9 @@
 import { type GetTagValuesRequest, pathEncode } from "@vantage-sh/vantage-client";
 import z from "zod";
-import paginationData from "../utils/paginationData";
-import { DEFAULT_LIMIT } from "./structure/constants";
-import MCPUserError from "./structure/MCPUserError";
-import registerTool from "./structure/registerTool";
+import paginationData from "../../utils/paginationData";
+import { DEFAULT_LIMIT } from "../structure/constants";
+import MCPUserError from "../structure/MCPUserError";
+import registerTool from "../structure/registerTool";
 
 const description = `
 List values for a tag key. The argument is \`key\` (the tag key name); the API response fields use \`tag_value\`.
