+++ src/tools/tags/list-tags.ts
@@ -1,9 +1,9 @@
 import type { GetTagsRequest } from "@vantage-sh/vantage-client";
 import z from "zod";
-import paginationData from "../utils/paginationData";
-import { DEFAULT_LIMIT } from "./structure/constants";
-import MCPUserError from "./structure/MCPUserError";
-import registerTool from "./structure/registerTool";
+import paginationData from "../../utils/paginationData";
+import { DEFAULT_LIMIT } from "../structure/constants";
+import MCPUserError from "../structure/MCPUserError";
+import registerTool from "../structure/registerTool";
 
 const description = `
 List tags that can be used to filter costs and cost reports.
