+++ test/tools/tags/list-tag-values.test.ts
@@ -1,7 +1,7 @@
-import type { GetTagsResponse } from "@vantage-sh/vantage-client";
+import { type GetTagValuesRequest, type GetTagValuesResponse, pathEncode } from "@vantage-sh/vantage-client";
 import { expect } from "vitest";
-import tool from "../../src/tools/list-tags";
-import { DEFAULT_LIMIT } from "../../src/tools/structure/constants";
+import tool from "../../../src/tools/tags/list-tag-values";
+import { DEFAULT_LIMIT } from "../../../src/tools/structure/constants";
 import {
   type ExecutionTestTableItem,
   type ExtractOutputSchema,
@@ -10,12 +10,13 @@ import {
   requestsInOrder,
   type SchemaTestTableItem,
   testTool,
-} from "../../src/utils/testing";
+} from "../../../src/utils/testing";
 
 type Validators = ExtractValidators<typeof tool>;
 type OutputSchema = ExtractOutputSchema<typeof tool>;
 
 const validArguments: InferValidators<Validators> = {
+  key: "environment",
   page: 1,
   search_query: undefined,
   providers: undefined,
@@ -23,24 +24,26 @@ const validArguments: InferValidators<Validators> = {
 
 const argumentSchemaTests: SchemaTestTableItem<Validators>[] = [
   {
-    name: "default page",
+    name: "blank string key",
     data: {
-      page: undefined,
+      key: "",
+      page: 1,
       search_query: undefined,
       providers: undefined,
     },
+    expectedIssues: ["Too small: expected string to have >=1 characters"],
   },
   {
-    name: "valid page number",
+    name: "valid arguments",
     data: validArguments,
   },
 ];
 
-const successData: GetTagsResponse = {
-  tags: [
-    { tag_key: "environment", hidden: false, preferred: false, providers: ["aws", "azure"] },
-    { tag_key: "project", hidden: false, preferred: false, providers: ["aws", "gcp"] },
-    { tag_key: "team", hidden: false, preferred: false, providers: ["aws"] },
+const successData: GetTagValuesResponse = {
+  tag_values: [
+    { tag_value: "production", providers: ["aws", "azure"] },
+    { tag_value: "staging", providers: ["aws"] },
+    { tag_value: "development", providers: ["aws", "gcp"] },
   ],
   links: {},
 };
@@ -50,13 +53,11 @@ const executionTests: ExecutionTestTableItem<Validators, OutputSchema>[] = [
     name: "successful call",
     apiCallHandler: requestsInOrder([
       {
-        endpoint: "/v2/tags",
+        endpoint: `/v2/tags/${pathEncode(validArguments.key)}/values`,
         params: {
-          page: 1,
-          search_query: undefined,
-          providers: undefined,
+          ...validArguments,
           limit: DEFAULT_LIMIT,
-        },
+        } as GetTagValuesRequest,
         method: "GET",
         result: {
           ok: true,
@@ -67,7 +68,7 @@ const executionTests: ExecutionTestTableItem<Validators, OutputSchema>[] = [
     handler: async ({ callExpectingSuccess }) => {
       const res = await callExpectingSuccess(validArguments);
       expect(res).toEqual({
-        tags: successData.tags,
+        tag_values: successData.tag_values,
         pagination: {
           hasNextPage: false,
           nextPage: 0,
@@ -79,24 +80,23 @@ const executionTests: ExecutionTestTableItem<Validators, OutputSchema>[] = [
     name: "unsuccessful call",
     apiCallHandler: requestsInOrder([
       {
-        endpoint: "/v2/tags",
+        endpoint: `/v2/tags/${pathEncode(validArguments.key)}/values`,
         params: {
+          ...validArguments,
           page: 1,
-          search_query: undefined,
-          providers: undefined,
           limit: DEFAULT_LIMIT,
-        },
+        } as GetTagValuesRequest,
         method: "GET",
         result: {
           ok: false,
-          errors: [{ message: "Access denied" }],
+          errors: [{ message: "Tag not found" }],
         },
       },
     ]),
     handler: async ({ callExpectingMCPUserError }) => {
       const err = await callExpectingMCPUserError(validArguments);
       expect(err.exception).toEqual({
-        errors: [{ message: "Access denied" }],
+        errors: [{ message: "Tag not found" }],
       });
     },
   },
